Fixed issue with mount/unmount Skia views.
A fix was added to the PR that unregistered the SkiaView when it was unmounted on navigation.

We now go back to just set it to nil to free the draw view.

We also use a weakPtr to subscribe to notifications, and I also readded back dealloc with unregistration of the skia view.
Also added unsubscribing to the notification in dealloc.