Added reset method to Matrix to avoid recreating Matrix objects
When performance is an issue recreating a lot of JsiSkXXX objects when rendering each frame can bring the performance down.

This is caused by the translation layer between C++/JSI and Javascript where new objects needs to look up functions from names the first time a function or property is called on the object.

This PR adds a reset method on the Matrix object (both Web/Native) so that we are able to reuse Matrix objects when creating performance heavy renderings. This lets us use Matrices directly in Group tags without having to parse / read the transform on each render - and without having to create new objects on each render.