Refactored ViewAPI to be more generic
To be able to implement different ways of rendering with Skia (different subclasses of RNSkiaViews) we need a more generic View API for communication between JS and C++.

Previously the SkiaView in JS sent the draw callback using the method `setDrawCallback` in the View API (to be able to set the function directly in C++ without serialization).

If we were to create a Skia View for rendering SkPictures only we don't want to set a draw callback but we want to be able to set a picture property.

With this change the ViewAPI now exposes two generic methods, one for setting props and one for calling methods (actions):

- setCustomProp
- callCustomAction

(In addition the registerValuesInView method is kept since it can be called without a view instance). All previous methods have been removed.

This means that we now have a more flexible and generic API for Skia views that can be used to build future Skia views that requires/uses other properties than the draw callback.

This PR also simplifies storing the draw callback in the draw view.