Removing use of PointerEvents.canBeTouchTarget
This function was introduced in React Native 0.68 and should not be used if we want to keep our backwards compatibility on Android.

Luckily this call isn't necessary to implement / support the pointer events example anyways, so it can be removed.

This commit removes this.

Fixes Cannot build on Android using 0.67.5 due to Skia using newer APIs. #1216