An extra std::move nulls out the context param
When creating a JS based renderer we are passing the platform context - the ctor is using std::move to move the reference which causes use of the context later to be nullptr - resulting in nullptr being returned when trying to get the platform context from within the renderer scope.

This can be seen if you try to access the platform context from within any JsiSkCanvas method in a sample using a SkiaViews.