iOS: Reverted commit to fix memory regression
We changed the setup of the Skia Metal backend to use SkSurface::MakeFromCAMetalLayer. This encapsluates the nextDrawable call.

We previously had control over this inside our autoreleasepool, but this change removes this control and makes memory grow and layers not being released.

I reverted this code back and re-measured everything and then the memory leak was gone.

Tested with release mode on device.